Reorganized a bit
